CourseSearchBot/requirements.txt
pdfplumber==0.11.4
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy==1.26.4
fpdf==1.7.2
PyYAML==6.0.2
Pillow==10.4.0
# Optional — uncomment if you want OCR for scanned PDFs:
# pytesseract==0.3.13
